% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{att}
\alias{att}
\title{Extract filtered state estimates used in KFS}
\usage{
att(object)
}
\arguments{
\item{object}{KFS object}
}
\description{
Accessor method to access filtered state estimate in KFS object
}
