% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid}
\alias{covid}
\title{Cumulative cases of Covid-19 in Italy and England, before 14 Dec 2020.}
\format{
An object of class `"xts"`;
\describe{
  \item{Cases}{Cumulative cases of Covid-19}
}
}
\usage{
data(covid)
}
\description{
Cumulative cases of Covid-19 in Italy and England, before 14 Dec 2020.
}
\examples{
data(england)
# plot daily cases
plot(diff(england))
}
\references{
Downloaded from https://ukhsa-dashboard.data.gov.uk/topics/covid-19
}
\keyword{datasets}
