% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{gety.hat}
\alias{gety.hat}
\title{Extract prediction y.hat in SSModel}
\usage{
gety.hat(object)
}
\arguments{
\item{object}{SSModel object}
}
\description{
Accessor method to access time series y in SSModel object
}
