% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mape}
\alias{mape}
\title{Returns Mean Absolute Percentage Error (MAPE) of forecast compared to
a holdout sample}
\usage{
mape(res, n.ahead, Y)
}
\arguments{
\item{res}{filterResults or filterResultsLI object obtained from SSModel
functions.}

\item{n.ahead}{The number of days to forecast ahead.}

\item{Y}{The original cumulated dataset with 2 columns: leading indicator
and target variable.}
}
\value{
List object containing Mean Absolute Percentage Error (MAPE) of
trend forecast and forecast with seasonal component.
}
\description{
Helper method to compute the Mean Absolute Percentage Error
(MAPE) of a forecast generated by tsgc models compared to a holdout sample.
}
\examples{
eng = england[, c("cum_cases","cum_admissions")]
out<-SSModelLeadingIndicator(Y=eng,n.lag = 7)
res<-out$get_model()
mape(res,n.forecasts,Y)$sea

}
