% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{seasonalComp}
\alias{seasonalComp}
\title{Extract number of seasonal components used in KFS}
\usage{
seasonalComp(object)
}
\arguments{
\item{object}{KFS object}
}
\description{
Accessor method to access the number of seasonal component used in KFS object
}
