% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{matrixKFS}
\alias{matrixKFS}
\title{Extract matrices used in observation, state and disturbance equation
in KFS object}
\usage{
matrixKFS(object, matrix)
}
\arguments{
\item{object}{KFS object}

\item{matrix}{Any matrix component of SSModel, e.g. H,T,R,Q}
}
\description{
Accessor method to access the matrices used in observation,
state and disturbance equation in KFS object
}
